Redo notification helper and remove unused font
