+++ packages/app/src/screens/index.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
+import { Notifications } from 'expo';
+import { Notification } from 'expo/build/Notifications/Notifications.types';
 import NavigationService from './NavigationService';
 import Onboarding from './Onboarding';
 import Main from './Main';
@@ -18,6 +20,24 @@ const AppNavigator = ({ loggedIn }: AppNavigatorProps) => {
 			}
 		)
 	);
+
+	const handleNotification = (notification: Notification) => {
+		if (notification && notification.origin === 'selected') {
+			NavigationService.navigate('EmergencyDetails', {
+				details: notification.data
+			});
+		}
+	};
+
+	React.useEffect(() => {
+		const notificationSubscription = Notifications.addListener(
+			handleNotification
+		);
+		return () => {
+			notificationSubscription.remove();
+		};
+	}, []);
+
 	return (
 		<Navigator
 			ref={navigatorRef => {
